+++ omechu-app/src/app/auth/forgot-password/components/ForgotPasswordForm.tsx
@@ -1,9 +1,6 @@
 "use client";
 
-import { useState } from "react";
-
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
@@ -12,13 +9,16 @@ import Input from "@/app/auth/components/Input";
 import SquareButton from "@/app/components/common/button/SquareButton";
 import {
   findPasswordSchema,
-  FindPasswordFormValues,
+  type FindPasswordFormValues,
 } from "@/lib/schemas/auth.schema";
 
-export default function FindPasswordPage() {
-  const [apiError, setApiError] = useState<string | null>(null);
-  const router = useRouter();
+type ForgotPasswordFormProps = {
+  onFormSubmit: (data: FindPasswordFormValues) => Promise<void>;
+};
 
+export default function ForgotPasswordForm({
+  onFormSubmit,
+}: ForgotPasswordFormProps) {
   const {
     register,
     handleSubmit,
@@ -27,33 +27,10 @@ export default function FindPasswordPage() {
     resolver: zodResolver(findPasswordSchema),
   });
 
-  const onSubmit = async (data: FindPasswordFormValues) => {
-    setApiError(null);
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-    console.log(data);
-
-    if (data.email === "test@naver.com") {
-      router.push("/auth/find-password/sent");
-    } else {
-      setApiError(
-        "비밀번호 재설정 메일 발송에 실패했습니다. \n 이메일 주소를 다시 확인해 주세요.",
-      );
-    }
-  };
-
   return (
-    <div className="flex w-full flex-col items-center gap-6 pb-40">
-      <div className="flex flex-col gap-3 text-center">
-        <h1 className="text-xl font-medium text-[#393939]">비밀번호 찾기</h1>
-        <p className="text-sm font-normal text-[#828282]">
-          가입하신 이메일 주소를 입력하여
-          <br />
-          비밀번호를 재설정하실 수 있어요
-        </p>
-      </div>
-
+    <>
       <form
-        onSubmit={handleSubmit(onSubmit)}
+        onSubmit={handleSubmit(onFormSubmit)}
         className="flex w-full flex-col gap-2 pt-4"
       >
         <Input
@@ -77,31 +54,25 @@ export default function FindPasswordPage() {
       </form>
 
       <div className="flex flex-col items-center gap-1 pt-1 text-[13px] text-[#828282]">
-        <div className="flex items-center gap-4">
+        <div className="flex items-center justify-center gap-4">
           <span>비밀번호가 생각났어요</span>
           <Link
-            href="/auth/login"
+            href="/auth/sign-in"
             className="font-semibold text-[#494949] hover:underline"
           >
             로그인 하기
           </Link>
         </div>
-        <div className="flex items-center gap-4">
+        <div className="flex items-center justify-center gap-4">
           <span>계정이 아직 없어요</span>
           <Link
-            href="/auth/signup"
+            href="/auth/sign-up"
             className="font-semibold text-[#494949] hover:underline"
           >
             회원가입하기
           </Link>
         </div>
       </div>
-
-      {apiError && (
-        <div className="mt-4 flex h-[70px] w-full items-center justify-center whitespace-pre-line rounded-md bg-[rgba(130,130,130,0.5)] text-center text-sm text-white">
-          {apiError}
-        </div>
-      )}
-    </div>
+    </>
   );
 }
