+++ omechu-app/src/app/auth/forgot-password/sent/components/EmailSentMessage.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 
 import RoundButton from "@/app/components/common/button/RoundButton";
 
-export default function FindPasswordEmailSentPage() {
+export default function EmailSentMessage() {
   return (
     <div className="flex w-full flex-col items-center gap-10 pb-40 text-center">
       <div className="flex flex-col gap-3">
@@ -18,7 +18,7 @@ export default function FindPasswordEmailSentPage() {
         </p>
       </div>
       <div className="w-full px-24">
-        <Link href="/auth/login" className="w-full">
+        <Link href="/auth/sign-in" className="w-full">
           <RoundButton variant="red" size="lg" className="w-full">
             로그인 하기
           </RoundButton>
@@ -26,4 +26,4 @@ export default function FindPasswordEmailSentPage() {
       </div>
     </div>
   );
-}
+}
