+++ omechu-app/src/app/auth/forgot-password/sent/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import EmailSentMessage from "./components/EmailSentMessage";
+
+export default function FindPasswordEmailSentPage() {
+  return <EmailSentMessage />;
+}
